refactor exercise link setup
