+++ app/components/ExerciseDisplay.tsx
