--- app/exercises/toast-component-starting-styles/metadata.ts
@@ -1,10 +1,0 @@
-import { ExerciseMetadata } from "../../lib/exercises";
-
-export const metadata: ExerciseMetadata = {
-  title: "Transitions - Toast Component [With CSS Starting Styles]",
-  description:
-    "A toast notification system with stacking animations and smooth transitions",
-  courseUrl: "https://animations.dev/learn/css-animations/transitions",
-  topic: "CSS Animations",
-  slug: "toast-component-starting-styles",
-};
